reporter/requirements.txt
CherryPy==18.8.0
huggingface_hub==0.26.2
python-dotenv==1.0.1
reportlab==4.2.5
Requests==2.32.3
